10408: split getBulkTrialSessionWorkingCopies to separate file
