+++ web-api/src/persistence/dynamo/trialSessions/getTrialSessionWorkingCopy.test.ts
@@ -1,14 +1,10 @@
 import { applicationContext } from '../../../../../shared/src/business/test/createTestApplicationContext';
-import { batchGet, get } from '../../dynamodbClientService';
-import {
-  getBulkTrialSessionWorkingCopies,
-  getTrialSessionWorkingCopy,
-} from './getTrialSessionWorkingCopy';
+import { get } from '../../dynamodbClientService';
+import { getTrialSessionWorkingCopy } from './getTrialSessionWorkingCopy';
 
 jest.mock('../../dynamodbClientService');
 
 const getMock = get as jest.Mock;
-const batchGetMock = batchGet as jest.Mock;
 
 getMock.mockReturnValue({
   pk: 'trial-session-working-copy|123',
@@ -18,63 +14,6 @@ getMock.mockReturnValue({
   trialSessionId: '123',
   userId: '456',
 });
-
-batchGetMock.mockReturnValue([
-  {
-    caseMetadata: {},
-    entityName: 'TrialSessionWorkingCopy',
-    filters: {
-      basisReached: true,
-      continued: true,
-      definiteTrial: true,
-      dismissed: true,
-      motionToDismiss: true,
-      probableSettlement: true,
-      probableTrial: true,
-      recall: true,
-      rule122: true,
-      setForTrial: true,
-      settled: true,
-      showAll: true,
-      statusUnassigned: true,
-      submittedCAV: true,
-    },
-    pk: 'trial-session-working-copy|111ac21b-99f9-4321-98c8-b95db00af96b',
-    sessionNotes: 'Judge Colvin Super notes!',
-    sk: 'user|dabbad00-18d0-43ec-bafb-654e83405416',
-    sort: 'docket',
-    sortOrder: 'asc',
-    trialSessionId: '111ac21b-99f9-4321-98c8-b95db00af96b',
-    userId: 'dabbad00-18d0-43ec-bafb-654e83405416',
-  },
-  {
-    caseMetadata: {},
-    entityName: 'TrialSessionWorkingCopy',
-    filters: {
-      basisReached: true,
-      continued: true,
-      definiteTrial: true,
-      dismissed: true,
-      motionToDismiss: true,
-      probableSettlement: true,
-      probableTrial: true,
-      recall: true,
-      rule122: true,
-      setForTrial: true,
-      settled: true,
-      showAll: true,
-      statusUnassigned: true,
-      submittedCAV: true,
-    },
-    pk: 'trial-session-working-copy|0d943468-bc2e-4631-84e3-b084cf5b1fbb',
-    sessionNotes: 'Cohen Cohen Cohen Notes',
-    sk: 'user|dabbad04-18d0-43ec-bafb-654e83405416',
-    sort: 'docket',
-    sortOrder: 'asc',
-    trialSessionId: '0d943468-bc2e-4631-84e3-b084cf5b1fbb',
-    userId: 'dabbad04-18d0-43ec-bafb-654e83405416',
-  },
-]);
 
 describe('getTrialSessionWorkingCopy', () => {
   it('should get the trial session by id', async () => {
@@ -90,34 +29,6 @@ describe('getTrialSessionWorkingCopy', () => {
       sortOrder: 'desc',
       trialSessionId: '123',
       userId: '456',
-    });
-  });
-});
-
-describe('getBulkTrialSessionWorkingCopies', () => {
-  it('should get the trial session notes by special session array', async () => {
-    const specialTrialSessions = [
-      { pk: '123', sk: '456' },
-      { pk: '456', sk: '789' },
-    ];
-
-    const result = await getBulkTrialSessionWorkingCopies({
-      applicationContext,
-      specialTrialSessions,
-    });
-    expect(result).toEqual([
-      {
-        sessionNotes: 'Judge Colvin Super notes!',
-        trialSessionId: '111ac21b-99f9-4321-98c8-b95db00af96b',
-      },
-      {
-        sessionNotes: 'Cohen Cohen Cohen Notes',
-        trialSessionId: '0d943468-bc2e-4631-84e3-b084cf5b1fbb',
-      },
-    ]);
-    expect(batchGetMock).toHaveBeenCalledWith({
-      applicationContext,
-      keys: specialTrialSessions,
     });
   });
 });
